factored search results out of app.js
