+++ controllers/userProfileSetUp.js
@@ -1,6 +1,5 @@
 const mysql = require('../dbcon.js');
 const utils = require('../utils.js');
-const jp = require("jsonpath");
 
 module.exports = {
     getUserProfile: (req, res, next) => {
